More methods in DatabaseService and User

+++ src/main/java/edu/umb/cs/entity/Game.java
@@ -77,7 +77,7 @@ public class Game
         return puzzle;
     }
     
-    public int getPoint()
+    public long getPoint()
     {
         return point;
     }
